+++ src/app/feat-recruit/interactions/fest_recruit.ts
@@ -1,7 +1,6 @@
 import { Role } from '@prisma/client';
 import { ChatInputCommandInteraction, ModalSubmitInteraction } from 'discord.js';
 
-import { registerRecruitData } from './registerRecruitData';
 import { RecruitType } from '../../../db/recruit_service';
 import { RoleService } from '../../../db/role_service';
 import { getFesRegularData } from '../../common/apis/splatoon3.ink/splatoon3_ink';
@@ -11,6 +10,7 @@ import { RecruitOpCode } from '../canvases/regenerate_canvas';
 import { recruitAutoClose } from '../common/auto_close';
 import { arrangeRecruitData } from '../common/create_recruit/arrange_command_data';
 import { arrangeModalRecruitData } from '../common/create_recruit/arrange_modal_data';
+import { registerRecruitData } from '../common/create_recruit/register_recruit_data';
 import { removeDeleteButton } from '../common/create_recruit/remove_delete_button';
 import {
     sendRecruitCanvas,
