+++ src/app/feat-recruit/interactions/regular_recruit.ts
@@ -1,6 +1,5 @@
 import { ChatInputCommandInteraction, ModalSubmitInteraction } from 'discord.js';
 
-import { registerRecruitData } from './registerRecruitData';
 import { RecruitType } from '../../../db/recruit_service';
 import { UniqueRoleService } from '../../../db/unique_role_service';
 import { getRegularData } from '../../common/apis/splatoon3.ink/splatoon3_ink';
@@ -11,6 +10,7 @@ import { recruitRegularCanvas, ruleRegularCanvas } from '../canvases/regular_can
 import { recruitAutoClose } from '../common/auto_close';
 import { arrangeRecruitData } from '../common/create_recruit/arrange_command_data';
 import { arrangeModalRecruitData } from '../common/create_recruit/arrange_modal_data';
+import { registerRecruitData } from '../common/create_recruit/register_recruit_data';
 import { removeDeleteButton } from '../common/create_recruit/remove_delete_button';
 import {
     sendRecruitCanvas,
